role grid list & all fetch service does not have permission

+++ src/components/ResourceSetting/TrinityMembers/UserGroup/UserGroupMemberPanel.vue
@@ -1,8 +1,8 @@
 <template>
-    <div>
+    <div class="w3-small">
         <apply-members-window v-if="applyWindowAlive" :windowAlive="applyWindowAlive" 
-            window-title="Apply Members To User Group" 
             :memberUids="memberUids" 
+            :groupname="groupname" 
             :groupuid="groupuid"  
             @closeApply="changeApplyWindowStatus" 
             @applyMembers="getMembers" 
@@ -15,7 +15,7 @@
             @closeDelete="changeDeleteWindowStatus" 
             @confirmDelete="deleteApply" 
         ></confirm-delete-window>
-        <div class="w3-small">
+        <div v-if="applyMembers.length > 0">
             <div class="w3-row">
                 <div class="w3-col m10">
                     <input :id="'SearchUserInput' + groupuid" class="w3-input w3-border w3-border-camo-black w3-grey" type="text" 
@@ -43,7 +43,7 @@
     </div>
 </template>
 <script>
-import { HTTP_TRINITY,errorHandle } from '../../../../util_js/axios_util'
+import { HTTP_TRINITY,HTTP_AUTH,errorHandle } from '../../../../util_js/axios_util'
 import ApplyMembersWindow from './ApplyMembersWindow.vue'
 import ConfirmDeleteWindow from '../../ConfirmDeleteWindow.vue'
 
@@ -60,7 +60,8 @@ export default {
         }
     },
     props:{
-        groupuid: ''
+        groupuid: '',
+        groupname: ''
     },
     created(){
         this.getMembers()
@@ -82,22 +83,43 @@ export default {
         },
         getMembers(){
             if(this.groupuid && this.groupuid !== ''){
-                HTTP_TRINITY.get(`group-member/findFullNameByGroupUid`, {params:{uid: this.groupuid}})
+                HTTP_AUTH.get(`authorization/isRootOrAdmin`)
                 .then(response => {
-                    this.applyMembers = response.data
-                    this.memberUids = []
-                    for(let i=0; i<this.applyMembers.length; i++){
-                        this.memberUids.push(this.applyMembers[i].useruid)
+                    if(response.data){
+                        HTTP_TRINITY.get(`group-member/findFullNameByGroupUid`, {params:{uid: this.groupuid}})
+                        .then(response => {
+                            this.applyMembers = response.data
+                            this.memberUids = []
+                            for(let i=0; i<this.applyMembers.length; i++){
+                                this.memberUids.push(this.applyMembers[i].useruid)
+                            }
+                        })
+                        .catch(error => {
+                            errorHandle(this.$store, error)
+                        })
                     }
                 })
                 .catch(error => {
                     errorHandle(this.$store, error)
                 })
             }
-            
         },
         changeApplyWindowStatus(){
-            this.applyWindowAlive = !this.applyWindowAlive
+            HTTP_AUTH.get(`authorization/isRootOrAdmin`)
+            .then(response => {
+                if(response.data){
+                    this.applyWindowAlive = !this.applyWindowAlive
+                }else{
+                    let newStatus = {
+                        "msg": "You do not have 'Root Or Admin' Permission!",
+                        "status": "Warn"
+                    }
+                    this.$store.dispatch('setSystemStatus', newStatus)
+                }
+            })
+            .catch(error => {
+                errorHandle(this.$store, error)
+            })
         },
         changeDeleteWindowStatus(index, list_info){
             this.deleteWindowAlive = !this.deleteWindowAlive
@@ -122,12 +144,26 @@ export default {
             if(!this.deleteUid || this.deleteUid === '')
                 return
             
-            HTTP_TRINITY.get('group-member/deleteByPKUids?groupUid='+this.groupuid+'&userUid='+this.deleteUid)
+            HTTP_AUTH.get(`authorization/isRootOrAdmin`)
             .then(response => {
-                this.applyMembers.splice(this.deleteIndex, 1)
-                this.memberUids.splice(this.deleteIndex, 1)
-                
-                this.changeDeleteWindowStatus()
+                if(response.data){
+                    HTTP_TRINITY.get('group-member/deleteByPKUids?groupUid='+this.groupuid+'&userUid='+this.deleteUid)
+                    .then(response => {
+                        this.applyMembers.splice(this.deleteIndex, 1)
+                        this.memberUids.splice(this.deleteIndex, 1)
+                        
+                        this.changeDeleteWindowStatus()
+                    })
+                    .catch(error => {
+                        errorHandle(this.$store, error)
+                    })
+                }else{
+                    let newStatus = {
+                        "msg": "You do not have 'Root Or Admin' Permission!",
+                        "status": "Warn"
+                    }
+                    this.$store.dispatch('setSystemStatus', newStatus)
+                }
             })
             .catch(error => {
                 errorHandle(this.$store, error)
